+++ tests/test_imports/importa/importb/__init__.py
@@ -1,0 +1,5 @@
+import logging
+import typing as tp
+
+logger = logging.getLogger(__name__)
+