+++ tests/test_imports/importa/importb/dontimportme.py
@@ -1,0 +1,7 @@
+import logging
+import typing as tp
+
+logger = logging.getLogger(__name__)
+
+def sub(a: float, b: float) -> float:
+    return a-b