+++ tests/test_imports/importa/importb/importme.py
@@ -1,0 +1,10 @@
+import logging
+import typing as tp
+
+logger = logging.getLogger(__name__)
+
+__all__ = ['add', 'would_have_failed']
+
+def add(a: float, b: float) -> float:
+    return a+b
+